Added optional session to all API calls
An optional session parameter has been added to all API call to enable
the reuse of existing request sessions. This can be used to reduce the
overhead incurred from establishing HTTP/HTTPS connections, and to
allow the callers to take advantage of the features of the request
session.

The addition of the session parameter has been introduced to preserve
backwards compatibility with the existing API calls.